Split up massive cpp file
