+++ src/pegr/gensys/GensysLuaInterface.hpp
@@ -13,11 +13,19 @@ namespace pegr {
 namespace Gensys {
 namespace LI {
 
+
 /**
  * @brief Initializes. Sets up workspaces for adding new components,
  * genres, archetypes, etc.
  */
 void initialize();
+
+/**
+ * @brief Undoes initialize()
+ */
+void cleanup();
+
+//// SETUP ////
 
 /**
  * @brief Produces a typed primitive value from a Lua value.
@@ -85,20 +93,22 @@ Interm::Genre* parse_genre(int table_idx);
  */
 void stage_all();
 
-void cleanup();
-
 int add_component(lua_State* l);
 int edit_component(lua_State* l);
 int add_archetype(lua_State* l);
 int edit_archetype(lua_State* l);
-int find_archetype(lua_State* l);
 int add_genre(lua_State* l);
+
+//// RUNTIME ////
+
+// ...
+int find_archetype(lua_State* l);
 
 // MTI = metatable id
 extern const char* MTI_ARCHETYPE;
 extern const char* MTI_ENTITY;
 
-int entity_new(lua_State* l);
+int new_entity(lua_State* l);
 
 } // namespace LI
 } // namespace Gensys
