+++ src/pegr/gensys/GensysLuaInterfaceExecution.cpp
@@ -1,0 +1,47 @@
+#include "pegr/gensys/GensysLuaInterface.hpp"
+
+#include <stdexcept>
+#include <sstream>
+#include <cassert>
+#include <cstddef>
+#include <cstdint>
+#include <iostream>
+#include <vector>
+#include <map>
+
+#include "pegr/debug/DebugMacros.hpp"
+#include "pegr/logger/Logger.hpp"
+#include "pegr/gensys/Gensys.hpp"
+#include "pegr/script/ScriptHelper.hpp"
+
+namespace pegr {
+namespace Gensys {
+namespace LI {
+
+const char* MTI_ARCHETYPE = "pegr.Archetype";
+const char* MTI_ENTITY = "pegr.Entity";
+
+int find_archetype(lua_State* l) {
+    if (Gensys::get_global_state() != GlobalState::EXECUTABLE) {
+        luaL_error(l, "find_archetype is only available during execution");
+    }
+    std::size_t strlen;
+    const char* strdata = luaL_checklstring(l, 1, &strlen);
+    std::string key(strdata, strlen);
+    // TODO: resolve namespace issues in key
+    
+    /*
+    auto iter = m_archetypes.find(key);
+    if (iter == m_archetypes.end()) {
+        // ???
+    }
+    */
+}
+
+int new_entity(lua_State* l) {
+    
+}
+
+} // namespace LI
+} // namespace Gensys
+} // namespace pegr
